+++ src/isEqual.js
@@ -2,8 +2,34 @@ import isArray from './utility/isArray';
 import isObject from './utility/isObject';
 import multiArgs from './utility/multiArgs';
 
+const getComparer = (a) => {
+	let alt;
+
+	if (isObject(a)) {
+		alt = Object.keys(a).length;
+		return (b) => isObject(b) && alt === Object.keys(b).length;
+	}
+	if (isArray(a)) {
+		alt = a.length;
+		return (b) => isArray(b) && alt === b.length;
+	}
+	if (a instanceof Date) {
+		alt = a.getTime();
+		return (b) => b instanceof Date && Object.is(b.getTime(), alt);
+	}
+	if (a instanceof RegExp) {
+		alt = a + '';
+		return (b) => b instanceof RegExp && Object.is(b + '', alt);
+	}
+	if (a !== a) {
+		return (b) => b !== b;
+	}
+
+	return (b) => Object.is(b, a);
+};
+
 /**
- * Shallow compares two or more items. All items are compared with strict equality except Dates and RegExps which compare their _values_ with strict equality and Objects and Arrays which compare key lengths.
+ * Shallow compares two or more items. All items are compared with [SameValue](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Equality_comparisons_and_sameness#Same-value_equality) equality except Dates and RegExps which compare their _values_ with [SameValue](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Equality_comparisons_and_sameness#Same-value_equality) equality and Objects and Arrays which compare key lengths.
  *
  * @example
  * ``` javascript
@@ -27,41 +53,8 @@ import multiArgs from './utility/multiArgs';
  */
 export default (...args) => {
 	args = multiArgs(args);
-	const base = args.shift();
-	let baseType = '';
 
-	if (isObject(base)) {
-		baseType = 'Object';
-	}
-	else if (isArray(base)) {
-		baseType = 'Array';
-	}
-	else if (base instanceof Date) {
-		baseType = 'Date';
-	}
-	else if (base instanceof RegExp) {
-		baseType = 'RegExp';
-	}
-	else if (base !== base) {
-		baseType = 'NaN';
-	}
+	const base = args.pop();
 
-	return args.every((item) => {
-		if (baseType === 'Object') {
-			return isObject(item) && Object.keys(base).length === Object.keys(item).length;
-		}
-		if (baseType === 'Array') {
-			return isArray(item) && base.length === item.length;
-		}
-		if (baseType === 'NaN') {
-			return item !== item;
-		}
-		if (baseType === 'Date') {
-			return item instanceof Date && item.getTime() === base.getTime();
-		}
-		if (baseType === 'RegExp') {
-			return item instanceof RegExp && String(item) === String(base);
-		}
-		return item === base;
-	});
+	return args.every(getComparer(base));
 };